Show logged out empty observation screen before loading either MyObs screen
